Fix sub-supervisors in PID namespace
Supervisors can be chained:

    supervisor -> supervisor -> sh -> cat

The 2 supervisors may not be in the same PID namespace:

    supervisor (global) -> supervisor (namespace) -> sh -> cat

Code requesting the PID of sh using pidof/1 gets the PID in the namespace.

If a supervisor in the global PID namespace now attempts to signal that
PID, it will signal an unrelated PID in the global namespace, not the
the PID corresponding to the PID in the private namespace.

To fix this, the code could do one of the following:

* find all descendent PIDs of the supervisor in the global namespace

* query the direct parent supervisor of the process since it either created
  the namespace or is inside the namespace

    supervisor (global) -> supervisor (namespace) -> supervisor -> sh -> cat
    supervisor (global) -> supervisor (namespace) -> supervisor(namespace (nested)) -> sh -> cat

To implement the second option, modify the init callback to return a
list of supervisors.