supervisor chain: reset signal handlers
The supervisor process signal handler uses the "self pipe" trick:
the signal handler converts a received signal to a write to a pipe
fd. The supervisor traps all catchable signals by default because an
init process must:

* rebroadcast signals to child processes
* wait on child processes to avoid re-parenting exited processes to PID
  1 as zombies

If the signal handler pipe buffer fills up, the write returns immediately,
the signal is dropped and the signal handler file descriptor is
closed. The read end of the pipe returns an error which should force
the supervisor to exit.

The "pipe full buffer" test writes to the standard input of an inactive
child process. The pipe buffer between parent/child fd's fills up and
the parent receives a SIGPIPE signal for each attempted write.

The supervisor event handler:

* busy loops handling SIGPIPE

* other events are not serviced: the read of the signal pipe must handle
  the buffered writes before returning an error and exiting

* the supervisor process eventually receives SIGSEGV(!)

With a supervisor chain, all supervisors may eventually block attempting
to write to the stdout of the next process in the pipeline.

To workaround the issue:

* set the signal handler to ignore SIGPIPE: write failures return EPIPE

* set the default action (abort) for SIGSEGV

This change fixes the lingering Supervise processes after the test run
by aborting the process.

Additionally the upstream alcove dep should be fixed:

* better define the behaviour on pipe failure writes (either crash or
  unregister the signal handlers)

* investigate and fix the SIGSEGV

To distinguish `supervise(8)` processes from Stdio supervise processes,
pun on elixir Atoms by renaming the supervise process to Supervise.