signal: ignore signals before broadcasting
Clean up the signal handling:

If a signal was, e.g., sent to the process group (PID 0), the signal
would be re-received by the Supervise process causing an infinite loop
of signalling. Ignore the signal before resending.

Use the platform defined reap function to signal descendent processes.
For platforms without procfs or subreaper support, the Supervise process
can track PIDs and signal the process group (TODO).

Simplify by removing the untested/unused support for converting :EXIT
and :stream_signal messages to Unix signals. These can be re-added
later if useful: the idea was for an erlang process to send a message
to signal a process group. For example, SIGHUP to cause an application
to re-read configuration.